[cr137] [Android] Make AppMenuPropertiesDelegateImpl abstract
Chromium change:
https://chromium.googlesource.com/chromium/src/+/19050b6adac2e1f9b0442fc2dc4e58e5e023b251

Make AppMenuPropertiesDelegateImpl abstract.

Previously, the base class had all the logic for tabbed mode, and that
was never accessed by other child classes. This moves the tabbed logic
to the TabbedAppMenuPropertiesDelegate and make the base class abstract.

One of many CLs to try and remove XML menu inflation from the app menu.

Bug: 40145539
