+++ android/java/org/chromium/chrome/browser/app/appmenu/BraveAppMenuPropertiesDelegateImpl.java
@@ -6,12 +6,9 @@
 package org.chromium.chrome.browser.app.appmenu;
 
 import android.content.Context;
-import android.view.Menu;
-import android.view.MenuItem;
 import android.view.View;
 
 import androidx.annotation.Nullable;
-import androidx.appcompat.content.res.AppCompatResources;
 
 import org.chromium.base.supplier.ObservableSupplier;
 import org.chromium.base.supplier.OneshotSupplier;
@@ -19,9 +16,7 @@ import org.chromium.base.supplier.Supplier;
 import org.chromium.chrome.R;
 import org.chromium.chrome.browser.ActivityTabProvider;
 import org.chromium.chrome.browser.bookmarks.BookmarkModel;
-import org.chromium.chrome.browser.incognito.reauth.IncognitoReauthController;
 import org.chromium.chrome.browser.layouts.LayoutStateProvider;
-import org.chromium.chrome.browser.multiwindow.BraveMultiWindowUtils;
 import org.chromium.chrome.browser.multiwindow.MultiWindowModeStateDispatcher;
 import org.chromium.chrome.browser.readaloud.ReadAloudController;
 import org.chromium.chrome.browser.tabmodel.TabModelSelector;
@@ -32,7 +27,7 @@ import org.chromium.chrome.browser.ui.appmenu.AppMenuItemProperties;
 import org.chromium.ui.modelutil.MVCListAdapter;
 import org.chromium.ui.modelutil.MVCListAdapter.ModelList;
 
-public class BraveAppMenuPropertiesDelegateImpl extends AppMenuPropertiesDelegateImpl {
+public abstract class BraveAppMenuPropertiesDelegateImpl extends AppMenuPropertiesDelegateImpl {
     public BraveAppMenuPropertiesDelegateImpl(
             Context context,
             ActivityTabProvider activityTabProvider,
@@ -42,9 +37,6 @@ public class BraveAppMenuPropertiesDelegateImpl extends AppMenuPropertiesDelegat
             View decorView,
             @Nullable OneshotSupplier<LayoutStateProvider> layoutStateProvidersSupplier,
             ObservableSupplier<BookmarkModel> bookmarkModelSupplier,
-            @Nullable
-                    OneshotSupplier<IncognitoReauthController>
-                            incognitoReauthControllerOneshotSupplier,
             @Nullable Supplier<ReadAloudController> readAloudControllerSupplier) {
         super(
                 context,
@@ -55,15 +47,7 @@ public class BraveAppMenuPropertiesDelegateImpl extends AppMenuPropertiesDelegat
                 decorView,
                 layoutStateProvidersSupplier,
                 bookmarkModelSupplier,
-                incognitoReauthControllerOneshotSupplier,
                 readAloudControllerSupplier);
-    }
-
-    @Override
-    public void prepareMenu(Menu menu, AppMenuHandler handler) {
-        super.prepareMenu(menu, handler);
-
-        maybeReplaceIcons(menu);
     }
 
     @Override
@@ -89,48 +73,5 @@ public class BraveAppMenuPropertiesDelegateImpl extends AppMenuPropertiesDelegat
         }
 
         return modelList;
-    }
-
-    private void maybeReplaceIcons(Menu menu) {
-        if (shouldShowIconBeforeItem()) {
-            for (int i = 0; i < menu.size(); ++i) {
-                MenuItem item = menu.getItem(i);
-                if (item.hasSubMenu()) {
-                    maybeReplaceIcons(item.getSubMenu());
-                } else if (item.getItemId() == R.id.new_tab_menu_id) {
-                    item.setIcon(
-                            AppCompatResources.getDrawable(mContext, R.drawable.ic_new_tab_page));
-                } else if (item.getItemId() == R.id.new_incognito_tab_menu_id) {
-                    item.setIcon(AppCompatResources.getDrawable(
-                            mContext, R.drawable.brave_menu_new_private_tab));
-                } else if (item.getItemId() == R.id.all_bookmarks_menu_id) {
-                    item.setIcon(AppCompatResources.getDrawable(
-                            mContext, R.drawable.brave_menu_bookmarks));
-                } else if (item.getItemId() == R.id.recent_tabs_menu_id) {
-                    item.setIcon(AppCompatResources.getDrawable(
-                            mContext, R.drawable.brave_menu_recent_tabs));
-                } else if (item.getItemId() == R.id.open_history_menu_id) {
-                    item.setIcon(AppCompatResources.getDrawable(
-                            mContext, R.drawable.brave_menu_history));
-                } else if (item.getItemId() == R.id.downloads_menu_id) {
-                    item.setIcon(AppCompatResources.getDrawable(
-                            mContext, R.drawable.brave_menu_downloads));
-                } else if (item.getItemId() == R.id.preferences_id) {
-                    item.setIcon(AppCompatResources.getDrawable(
-                            mContext, R.drawable.brave_menu_settings));
-                }
-            }
-        }
-    }
-
-    @Override
-    protected boolean shouldShowNewWindow() {
-        return BraveMultiWindowUtils.shouldEnableMultiWindows() && super.shouldShowNewWindow();
-    }
-
-    @Override
-    protected boolean shouldShowMoveToOtherWindow() {
-        return BraveMultiWindowUtils.shouldEnableMultiWindows()
-                && super.shouldShowMoveToOtherWindow();
     }
 }
